third_party/WebKit/LayoutTests/fast/events/keyboard-event-source-device-event-sender-expected.txt

Test that keyboard events should have a sourceDevice and its firesTouchEvents should be false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


keydown
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
keypress
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
keyup
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
PASS successfullyParsed is true

TEST COMPLETE
